requirements.txt
opencv-python
numpy
pytesseract
streamlit
Pillow